pyproject.toml
[tool.poetry]
name = "djchat"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.3"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
drf-spectacular = "^0.26.3"
pillow = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
